Consolidate meshIO code into meshio.cpp
